Shade UI Authentication and IRB-mandated consent Acceptance (#17)
* Addition of frontend auth using amplify aws

* Working new aws auth with amplify, cognito, lambda. Passes on username and
user.sub (guid) from cognito to the existing user data as password (so that the
existing app json believes they have a unique password to match, which they do).
Moved .eslintrc from json to a js file for better support for comments and
dangling commas. Dockerfile now updates npm since old version of npm was
throwing warnings. Added supported node engines information to package.json;
that is, the version the dockerfile uses and the minimum version we recommend
using to run the UI app both in dev and prod.

* Removed connection form since server settings will not be
exposed, and user/password are handled by new auth.

* Added SHADE tag to amplify backend tags.

* Added consent page.

* Adding api with consent/ endpoint. consent.jsx will call this when user accepts terms.

* Refreshing session once user has accepted consent/terms in order to allow users
to continue in the app.

* Changed/added email subject/body to auth config.

* Generates unique username during registration

* Added aws-exports files for dev and prod to ./web/doc folder, including notes on how to build and run app.

* Adding build cp script to copy aws exports file. Needs to be copied over before building front-end bundle.